+++ src/MaaWpfGui/Services/Web/IHttpService.cs
@@ -17,7 +17,7 @@ using System.IO;
 using System.Net.Http;
 using System.Threading.Tasks;
 
-namespace MaaWpfGui.Services
+namespace MaaWpfGui.Services.Web
 {
     public interface IHttpService
     {
