+++ src/MaaWpfGui/ViewModels/UI/CopilotViewModel.cs
@@ -24,6 +24,7 @@ using MaaWpfGui.Constants;
 using MaaWpfGui.Helper;
 using MaaWpfGui.Main;
 using MaaWpfGui.Services;
+using MaaWpfGui.Services.Web;
 using Microsoft.Win32;
 using Newtonsoft.Json;
 using Newtonsoft.Json.Linq;
